add nil checks to replace_all_granules rake
My local db was throwing errors on a few collections. The checks added
here cover the problems I was having, likely due to my local being in a
weird state.

We’ll want to find a decent way to test this with a mirror of the
production database as I can’t point to exactly why some of my
collection were throwing errors. It was likely just old/bad data but I
can’t be sure.