CMRARC-418 - Testing revealed a few issues.  1) If the user is a (#57)
* CMRARC-418 - Testing revealed a few issues.  1) If the user is a
daac_curator, an associated daac should be assigned.  Modified the
ACL code to retrieve the daac associated with the daac curator and both are
now assigned to the user.   Note it uses the provider as the daac.

2) Added tests for pulling in daac curators.   Also if it is a daac
added a test to make sure by checking the page you log into, daac curators
don't have a section called "Unreviewed" only arc curators and admins.
Tested for this for their use cases as well.

3) omniauth-urs code was updated in the developer branch and for the gem was being cached, so had to rebuild by clearing the cache on the build machine.

* CMRARC-418 - code either returns nil or some string, so simplified the
if statement.

* CMRARC-418
* changed wording regarding if a user is not provisioned.
* pulls provider_id from acl instead of parsing
* improved testings, looking for each section per the type of role user is provisioned
with.